pyproject.toml
[tool.poetry]
name = "seattle_fires_challenge_gc"
version = "0.1.0"
description = "Data Engineering Pipeline using Seattle Fires API"
authors = ["GoncaloCanteiro <gcanteiro15@gmail.com>"]
readme = "README.md"
packages = [{include = "seattle_fires"}]

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "3.4.1"
delta-spark = "^2.4.0"
ruff = "^0.0.284"
isort = "^5.12.0"
pytest = "^7.4.0"
chispa = "^0.9.2"


[tool.poetry.scripts]
bronze = "seattle_fires.tasks.ProcessBronze:process_bronze"
silver = "seattle_fires.tasks.ProcessSilver:process_silver"
gold = "seattle_fires.tasks.ProcessGold:process_gold"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

